# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Multi-Tenant SaaS POS System API — Phase 1 (Foundation)
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /suppliers:
    get:
      operationId: listSuppliers
      tags: [suppliers]
      summary: List suppliers
      parameters:
        - name: page
          in: query
          schema: { type: integer, minimum: 1, default: 1 }
        - name: pageSize
          in: query
          schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
        - name: search
          in: query
          schema: { type: string }
        - name: includeInactive
          in: query
          schema: { type: boolean, default: false }
        - name: withDebtOnly
          in: query
          schema: { type: boolean, default: false }
      responses:
        "200":
          description: Paginated suppliers
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SupplierList"
    post:
      operationId: createSupplier
      tags: [suppliers]
      summary: Create a supplier
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SupplierInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Supplier"
        "400":
          $ref: "#/components/responses/ValidationError"
        "409":
          $ref: "#/components/responses/Conflict"
  /suppliers/{id}:
    get:
      operationId: getSupplier
      tags: [suppliers]
      summary: Get a supplier
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: Supplier
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Supplier"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      operationId: updateSupplier
      tags: [suppliers]
      summary: Update a supplier
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SupplierInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Supplier"
        "400":
          $ref: "#/components/responses/ValidationError"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      operationId: deleteSupplier
      tags: [suppliers]
      summary: Deactivate a supplier
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "204":
          description: Deactivated
        "400":
          $ref: "#/components/responses/ValidationError"
        "404":
          $ref: "#/components/responses/NotFound"
  /suppliers/{id}/statement:
    get:
      operationId: getSupplierStatement
      tags: [suppliers]
      summary: Supplier ledger / statement
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: Statement
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SupplierStatement"
        "404":
          $ref: "#/components/responses/NotFound"
  /suppliers/{id}/payments:
    post:
      operationId: createSupplierPayment
      tags: [suppliers]
      summary: Record a supplier payment (reduces payable)
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentInput"
      responses:
        "201":
          description: Payment recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SupplierTransaction"
        "400":
          $ref: "#/components/responses/ValidationError"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  parameters:
    IdParam:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
  responses:
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Conflict:
      description: Conflict
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    Supplier:
      type: object
      required: [id, name, phone, currentBalance, isActive, createdAt]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        phone: { type: string }
        address: { type: ["string", "null"] }
        taxNumber: { type: ["string", "null"] }
        currentBalance: { type: string }
        notes: { type: ["string", "null"] }
        isActive: { type: boolean }
        createdAt: { type: string, format: date-time }
    SupplierInput:
      type: object
      required: [name, phone]
      properties:
        name: { type: string, minLength: 1 }
        phone: { type: string, minLength: 1 }
        address: { type: ["string", "null"] }
        taxNumber: { type: ["string", "null"] }
        notes: { type: ["string", "null"] }
    SupplierList:
      type: object
      required: [items, total, page, pageSize]
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/Supplier" }
        total: { type: integer }
        page: { type: integer }
        pageSize: { type: integer }
    SupplierTransaction:
      type: object
      required: [id, supplierId, type, debit, credit, balanceAfter, createdAt]
      properties:
        id: { type: string, format: uuid }
        supplierId: { type: string, format: uuid }
        type: { type: string }
        debit: { type: string }
        credit: { type: string }
        balanceAfter: { type: string }
        referenceType: { type: ["string", "null"] }
        referenceId: { type: ["string", "null"] }
        description: { type: ["string", "null"] }
        createdAt: { type: string, format: date-time }
    SupplierStatement:
      type: object
      required: [supplier, items, total, page, pageSize]
      properties:
        supplier: { $ref: "#/components/schemas/Supplier" }
        items:
          type: array
          items: { $ref: "#/components/schemas/SupplierTransaction" }
        total: { type: integer }
        page: { type: integer }
        pageSize: { type: integer }
    PaymentInput:
      type: object
      required: [amount, treasuryAccountId]
      properties:
        amount: { type: number, minimum: 0 }
        treasuryAccountId: { type: string, format: uuid }
        notes: { type: ["string", "null"] }
...